snap: Add dependencies to run under wayland
When running under wayland the gtk version embeded in the
snap will try to access to some xsettings provided by gnome
settings daemon, however there's no schema available and
this would read to a fatal error.

To prevent this, install the g-s-d schemas in the snap so that
the app could try to get such settings without crashing.